+++ server/faker-glx.cpp
@@ -1356,6 +1356,10 @@ void (*glXGetProcAddressARB(const GLubyte *procName))(void)
 		CHECK_FAKED(glFlush)
 		CHECK_FAKED(glDrawBuffer)
 		CHECK_FAKED(glDrawBuffers)
+		#ifdef GL_VERSION_4_5
+		CHECK_FAKED(glFramebufferDrawBufferEXT)
+		CHECK_FAKED(glFramebufferDrawBuffersEXT)
+		#endif
 		CHECK_FAKED(glGetBooleanv)
 		CHECK_FAKED(glGetDoublev)
 		CHECK_FAKED(glGetFloatv)
@@ -1363,6 +1367,10 @@ void (*glXGetProcAddressARB(const GLubyte *procName))(void)
 		CHECK_FAKED(glGetInteger64v)
 		CHECK_FAKED(glGetString)
 		CHECK_FAKED(glGetStringi)
+		#ifdef GL_VERSION_4_5
+		CHECK_FAKED(glNamedFramebufferDrawBuffer)
+		CHECK_FAKED(glNamedFramebufferDrawBuffers)
+		#endif
 		CHECK_FAKED(glPopAttrib)
 		CHECK_FAKED(glReadBuffer)
 		CHECK_FAKED(glViewport)