+++ server/faker-mapfile.c
@@ -97,6 +97,10 @@
 		glFlush;
 		glDrawBuffer;
 		glDrawBuffers;
+		#ifdef GL_VERSION_4_5
+		glFramebufferDrawBufferEXT;
+		glFramebufferDrawBuffersEXT;
+		#endif
 		glGetBooleanv;
 		glGetDoublev;
 		glGetFloatv;
@@ -104,6 +108,10 @@
 		glGetInteger64v;
 		glGetString;
 		glGetStringi;
+		#ifdef GL_VERSION_4_5
+		glNamedFramebufferDrawBuffer;
+		glNamedFramebufferDrawBuffers;
+		#endif
 		glPopAttrib;
 		glReadBuffer;
 		glViewport;